More operation on multi_action_name(_set) to more appropriate place.
