+++ libraries/process/include/mcrl2/process/alphabet_operations.h
@@ -21,141 +21,6 @@ namespace mcrl2::process::alphabet_operations
 {
 
 //-----------------------------------------------------//
-// multi_action_name operations
-//-----------------------------------------------------//
-
-// Returns true if the multiset y is contained in x
-inline
-bool includes(const multi_action_name& x, const multi_action_name& y)
-{
-  return std::includes(x.begin(), x.end(), y.begin(), y.end());
-}
-
-inline
-bool contains(const multi_action_name& alpha, const core::identifier_string& a)
-{
-  return alpha.find(a) != alpha.end();
-}
-
-// Returns alpha \ beta
-inline
-multi_action_name multiset_difference(const multi_action_name& alpha, const multi_action_name& beta)
-{
-  multi_action_name result = alpha;
-  for (const core::identifier_string& i: beta)
-  {
-    multi_action_name::iterator j = result.find(i);
-    if (j != result.end())
-    {
-      result.erase(j);
-    }
-  }
-  return result;
-}
-
-inline
-multi_action_name multiset_union(const multi_action_name& alpha, const multi_action_name& beta)
-{
-  multi_action_name result;
-  std::merge(alpha.begin(), alpha.end(), beta.begin(), beta.end(), std::inserter(result, result.end()));
-  return result;
-}
-
-//-----------------------------------------------------//
-// multi_action_name_set operations
-//-----------------------------------------------------//
-
-inline
-multi_action_name_set make_name_set(const action_name_multiset_list& v)
-{
-  multi_action_name_set result;
-  for (const action_name_multiset& i: v)
-  {
-    const core::identifier_string_list& names = i.names();
-    result.insert(multi_action_name(names.begin(), names.end()));
-  }
-  return result;
-}
-
-inline
-bool contains(const multi_action_name_set& A, const multi_action_name& a)
-{
-  // note that A implicitly contains the empty multi_action_name
-  return a.empty() || A.find(a) != A.end();
-}
-
-inline
-// Returns true if A contains an x such that includes(x, y)
-bool includes(const multi_action_name_set& A, const multi_action_name& y)
-{
-  // degenerate case
-  if (A.empty() && y.empty())
-  {
-    return true;
-  }
-  for (const multi_action_name& alpha: A)
-  {
-    if (includes(alpha, y))
-    {
-      return true;
-    }
-  }
-  return false;
-}
-
-inline
-bool subset_includes(const multi_action_name_set& A, const multi_action_name& x)
-{
-  for (const multi_action_name& a: A)
-  {
-    if (includes(a, x))
-    {
-      return true;
-    }
-  }
-  return false;
-}
-
-inline
-multi_action_name_set set_difference(const multi_action_name_set& A1, const multi_action_name_set& A2)
-{
-  multi_action_name_set result;
-  std::set_difference(A1.begin(), A1.end(), A2.begin(), A2.end(), std::inserter(result, result.end()));
-  return result;
-}
-
-inline
-multi_action_name_set set_union(const multi_action_name_set& A1, const multi_action_name_set& A2)
-{
-  multi_action_name_set result;
-  std::set_union(A1.begin(), A1.end(), A2.begin(), A2.end(), std::inserter(result, result.end()));
-  return result;
-}
-
-inline
-multi_action_name_set set_intersection(const multi_action_name_set& A1, const multi_action_name_set& A2)
-{
-  multi_action_name_set result;
-  std::set_intersection(A1.begin(), A1.end(), A2.begin(), A2.end(), std::inserter(result, result.end()));
-  return result;
-}
-
-// Removes elements of A that are a subset of another element.
-inline
-multi_action_name_set remove_subsets(const multi_action_name_set& A)
-{
-  multi_action_name_set result;
-  for (const multi_action_name& alpha: A)
-  {
-    if (!includes(result, alpha))
-    {
-      result.insert(alpha);
-    }
-  }
-  return result;
-}
-
-//-----------------------------------------------------//
 // block operations
 //-----------------------------------------------------//
 
@@ -457,7 +322,7 @@ inline
 void comm_inverse(const communication_expression_list& C, const multi_action_name& alpha1, const multi_action_name& alpha2, multi_action_name_set& result)
 {
   using utilities::detail::contains;
-  result.insert(alphabet_operations::multiset_union(alpha1, alpha2));
+  result.insert(multiset_union(alpha1, alpha2));
   for (const communication_expression& c: C)
   {
     if (contains(alpha1, c.name()))
